Move protocol event classes into a single file
